pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
authors = [
    {name = "Hieu", email = "hieu.buinguyen2352001@gmail.com"}
]
name = "chatbone"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.5.2",
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.13",
    "alembic>=1.15.1",
    "argon2-cffi>=23.1.0",
    "asyncpg>=0.30.0",
    "bitsandbytes>=0.45.4",
    "coverage>=7.6.12",
    "dramatiq[rabbitmq,redis,watch]>=1.17.1",
    "fastapi[standard]>=0.115.8",
    "huggingface>=0.0.1",
    "ipywidgets>=8.1.5",
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.3.8",
    "langgraph>=0.2.74",
    "langgraph-checkpoint-postgres>=2.0.19",
    "langgraph-cli[inmem]>=0.1.81",
    "loguru>=0.7.3",
    "notebook>=7.3.2",
    "psycopg[binary]>=3.2.6",
    "pwdlib[argon2]>=0.2.1",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "pytest-env>=1.1.5",
    "sqlalchemy[asyncio]>=2.0.38",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.49.0",
    "uuid7>=0.1.0",
    "uvloop>=0.21.0",
    "viztracer>=1.0.2",
]

[tool.setuptools]
#include-package-data = true # default

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope="session"
asyncio_default_test_loop_scope="session"
asyncio_mode = "strict"
testpaths="tests"
addopts="-v"


markers=[
    "unit", "auth", "utils","repo","workflows"
]
env=[
#    "LOGURU_LEVEL=INFO",
    "LOGURU_LEVEL=DEBUG",
]
